requirements.txt
appium-python-client==3.1.0
pytest==8.0.0
selenium==4.18.0
allure-pytest==2.13.2
pytest-xdist==3.5.0a